requirements.txt
ultralytics>=8.3.0,<9.0.0
opencv-python>=4.10.0,<5.0.0
numpy>=1.26.0,<3.0.0
PyYAML>=6.0.2,<7.0.0
pyttsx3>=2.98,<3.0.0
